refactor: extract shared StandardSourceQuery for view resolution
Dedupe RecursiveSourceQuery (query_executor.rs) and PrecomputeSourceQuery
(view_orchestrator.rs) into a single StandardSourceQuery in
src/fold_db_core/query/source_query.rs with a SourceQueryMode enum
(Recursive | Precompute) controlling how non-Cached views are handled.

- Recursive mode (user-facing query path): rejects views whose cache is
  in the Computing state, otherwise resolves through the view's inputs
  and persists Empty->Cached transitions.
- Precompute mode (background view orchestrator): treats any non-Cached
  state (including Computing) as Empty and inline-computes, relying on
  the orchestrator's bottom-up dependency walk.

Pure refactor: behavior is preserved exactly for both call sites.
