+++ src/fold_db_core/query/query_executor.rs
@@ -11,13 +11,13 @@ use crate::schema::types::Query;
 use crate::schema::SchemaError;
 use crate::schema::{SchemaCore, SchemaState};
 use crate::view::registry::ViewState;
-use crate::view::resolver::{SourceQueryFn, ViewResolver};
+use crate::view::resolver::ViewResolver;
 use crate::view::types::ViewCacheState;
-use async_trait::async_trait;
 use std::collections::HashMap;
 use std::sync::Arc;
 
 use super::hash_range_query::HashRangeQueryProcessor;
+use super::source_query::StandardSourceQuery;
 
 /// Main query executor that handles all query operations
 pub struct QueryExecutor {
@@ -25,96 +25,6 @@ pub struct QueryExecutor {
     db_ops: Arc<DbOperations>,
     hash_range_processor: HashRangeQueryProcessor,
     view_resolver: ViewResolver,
-}
-
-/// Implements SourceQueryFn by delegating back to the query executor's query path.
-/// This supports recursive resolution: views can query other views or schemas.
-struct RecursiveSourceQuery {
-    schema_manager: Arc<SchemaCore>,
-    db_ops: Arc<DbOperations>,
-    hash_range_processor: HashRangeQueryProcessor,
-    view_resolver: ViewResolver,
-}
-
-#[async_trait]
-impl SourceQueryFn for RecursiveSourceQuery {
-    async fn execute_query(
-        &self,
-        query: &Query,
-    ) -> Result<HashMap<String, HashMap<KeyValue, FieldValue>>, SchemaError> {
-        // First try as schema
-        match self.schema_manager.get_schema(&query.schema_name).await? {
-            Some(mut schema) => {
-                self.hash_range_processor
-                    .query_with_filter(
-                        &mut schema,
-                        &query.fields,
-                        query.filter.clone(),
-                        query.as_of,
-                    )
-                    .await
-            }
-            None => {
-                // Try as view (recursive)
-                self.try_query_view(query).await
-            }
-        }
-    }
-}
-
-impl RecursiveSourceQuery {
-    async fn try_query_view(
-        &self,
-        query: &Query,
-    ) -> Result<HashMap<String, HashMap<KeyValue, FieldValue>>, SchemaError> {
-        let view = {
-            let registry = self.schema_manager.view_registry().lock().map_err(|_| {
-                SchemaError::InvalidData("Failed to acquire view_registry lock".to_string())
-            })?;
-
-            registry
-                .get_view(&query.schema_name)
-                .cloned()
-                .ok_or_else(|| {
-                    SchemaError::NotFound(format!(
-                        "'{}' not found as schema or view",
-                        query.schema_name
-                    ))
-                })?
-        };
-
-        // Load cache state
-        let cache_state = self.db_ops.get_view_cache_state(&view.name).await?;
-
-        if cache_state.is_computing() {
-            return Err(SchemaError::InvalidData(format!(
-                "View '{}' is currently being precomputed and is not ready for queries",
-                view.name
-            )));
-        }
-
-        // Create a nested source query for this view's input queries
-        let nested_source = RecursiveSourceQuery {
-            schema_manager: Arc::clone(&self.schema_manager),
-            db_ops: Arc::clone(&self.db_ops),
-            hash_range_processor: HashRangeQueryProcessor::new(Arc::clone(&self.db_ops)),
-            view_resolver: ViewResolver::new(Arc::clone(self.view_resolver.wasm_engine())),
-        };
-
-        let (results, new_cache) = self
-            .view_resolver
-            .resolve(&view, &query.fields, &cache_state, &nested_source)
-            .await?;
-
-        // Persist cache if it changed from Empty to Cached
-        if cache_state.is_empty() && matches!(new_cache, ViewCacheState::Cached { .. }) {
-            self.db_ops
-                .set_view_cache_state(&view.name, &new_cache)
-                .await?;
-        }
-
-        Ok(results)
-    }
 }
 
 impl QueryExecutor {
@@ -297,12 +207,11 @@ impl QueryExecutor {
         }
 
         // Create source query implementation for recursive resolution
-        let source_query = RecursiveSourceQuery {
-            schema_manager: Arc::clone(&self.schema_manager),
-            db_ops: Arc::clone(&self.db_ops),
-            hash_range_processor: HashRangeQueryProcessor::new(Arc::clone(&self.db_ops)),
-            view_resolver: ViewResolver::new(Arc::clone(self.view_resolver.wasm_engine())),
-        };
+        let source_query = StandardSourceQuery::new_recursive(
+            Arc::clone(&self.schema_manager),
+            Arc::clone(&self.db_ops),
+            ViewResolver::new(Arc::clone(self.view_resolver.wasm_engine())),
+        );
 
         let (results, new_cache) = self
             .view_resolver
